Looker/requirements.txt
numpy==1.25.2
pandas==2.1.1
dash==2.11.1
plotly==5.17.0
google-cloud-bigquery==3.12.0